Add files via upload
In this project, we will build a MERN Stack application for the Chefkart platform, incorporating features inspired by popular food delivery apps to deliver a seamless and intuitive experience for users. By analyzing market-leading platforms, we will add functionalities and design elements that elevate the native look-and-feel of the app.